pyproject.toml
[tool.poetry]
name = "metrics"
version = "0.1.0"
description = ""
authors = ["91savage <ekdma1403@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
prometheus-client = "^0.17.0"
web3 = "^6.5.0"
setuptools = "^68.0.0"
python-dotenv = "^1.0.0"
substrate-interface = "^1.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
